Update view of list of groups
