--- src/app/groups/create/page.js
@@ -1,5 +1,0 @@
-export default function CreateGroup() {
-  return (
-    <h1>GROUP CREATE</h1>
-  )
-}
