--- src/app/groups/list/page.js
@@ -1,5 +1,0 @@
-export default function ListGroup() {
-  return (
-    <h1>GROUP LIST</h1>
-  )
-}
