+++ src/app/groups/loading.js
@@ -1,0 +1,5 @@
+'use client'
+
+export default function Loading() {
+  return <h1>Loading...</h1>
+}
