+++ src/app/services/groups.js
@@ -1,0 +1,10 @@
+import axios from "./axios";
+
+export const fetch = async () => {
+  try{
+    const response = await axios.get('/groups')
+    return response.data
+  }catch (error) {
+    throw error;
+  }
+}
